Addec cert viewer in external component.
