+++ src/pages/CertificateViewer.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import React, { useCallback, useState, useContext, useEffect } from 'react'
 import styled from 'styled-components'
 import { Grid } from '@digicatapult/ui-component-library'
@@ -15,236 +14,15 @@ import useAxios from '../hooks/use-axios'
 
 import CertificateViewHeader from './components/CertificateViewHeader'
 
-import CertificateTimeInterval from './components/CertificateTimeInterval'
+// import CertificateTimeInterval from './components/CertificateTimeInterval'
 
 import CertificateViewOwnership from './components/CertificateViewOwnership'
+import CertificateViewDetails from './components/CertificateViewDetails'
 
 // import LockerSvg from '../assets/images/locker-icon.svg'
 // import SealSvg from '../assets/images/approval-seal-small.svg'
 import BgMoleculesImageSVG from '../assets/images/molecules-bg-repeat.svg'
-import BgIconDateSVG from '../assets/images/icon-date.svg'
-import BgIconTimeSVG from '../assets/images/icon-time.svg'
-
-const PaddedWrapperDiv = styled.div`
-  padding: 4px;
-  height: calc(100% - 200px);
-  min-height: 400px;
-
-  text-align: left;
-
-  box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  -webkit-box-sizing: border-box;
-  border: 1px solid #111111;
-  background: #efefef;
-`
-
-const ContainerFlexWrapDiv = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-`
-
-const FlexDiv = styled.div`
-  margin: 0px 8px 20px 8px;
-  // height: 110px;
-  // line-height: 110px;
-  min-width: 169px;
-  flex-grow: 1;
-
-  color: #27847a;
-  border-left: 1px solid #27847a;
-
-  // background: #11cfcf;
-`
-
-const FlexRoundedDiv = styled.div`
-  margin: 0px 8px 20px 8px;
-  // height: 110px;
-  // line-height: 110px;
-  min-width: 169px;
-  flex-grow: 1;
-  padding: 20px;
-
-  border-radius: 10px;
-  background: #33e58c;
-`
-
-const FlexLargeDiv = styled(FlexDiv)`
-  min-width: 354px;
-
-  background: #d8d8d8;
-`
-
-const ContainerFlexNoWrapDiv = styled.div`
-  padding-left: 5px;
-  display: flex;
-  flex-wrap: nowrap;
-`
-
-const ContainerFullWidthWrapDiv = styled.div`
-  padding-left: 5px;
-  width: 100%;
-`
-
-const GrowingDiv = styled.div`
-  flex-grow: 1;
-`
-
-const HeadingDiv = styled.div`
-  // color: #1a1a1a;
-  // height: 28px;
-  // font: 700 15px/28px Roboto;
-  color: #1a1a1a;
-  height: 25px;
-  font: 700 15px/25px Roboto;
-`
-
-// const FixedWidthSmallDiv = styled.div`
-//   width: 98px;
-//   height: 82px;
-//   margin-top: 28px;
-//   text-align: center;
-//   color: #27847a;
-//   font: 500 14px/82px Roboto;
-//   border-left: 1px solid #27847a;
-//   border-right: 1px solid #27847a;
-//
-//   background: #d2d2d2;
-// `
-
-const InputWrapShortDiv = styled.div`
-  height: 42px;
-  line-height: 42px;
-  display: flex;
-  flex-wrap: nowrap;
-  padding: 15px 0 auto 0;
-  // padding-left: 15px;
-`
-
-const InputWrapDiv = styled.div`
-  height: 82px;
-  line-height: 82px;
-  display: flex;
-  flex-wrap: nowrap;
-  padding: 15px 0;
-`
-
-const InputHalfWrap = styled.div`
-  width: 50%;
-  height: 52px;
-  line-height: 52px;
-  display: flex;
-  flex-direction: column;
-  padding-left: 10px;
-`
-
-const IconWrap = styled.div`
-  display: flex;
-  width: 100%;
-  height: 26px;
-  font: 500 14px/26px Roboto;
-  color: #1a1a1a;
-`
-
-const IconDate = styled.span`
-  width: 26px;
-  height: 26px;
-  background: transparent url(${BgIconDateSVG}) no-repeat;
-`
-
-const IconTime = styled.span`
-  width: 26px;
-  height: 26px;
-  background: transparent url(${BgIconTimeSVG}) no-repeat;
-`
-
-const Text = styled.div`
-  display: flex;
-  width: 100%;
-  height: 26px;
-  min-width: 90px;
-  font: 500 18px/26px Roboto;
-  color: #1a1a1a;
-`
-
-const CertificateViewDetails = ({ size, start, end, energy, eco2 }) => {
-  return (
-    <>
-      <PaddedWrapperDiv>
-        <ContainerFlexWrapDiv>
-          {start && end && (
-            <FlexLargeDiv>
-              <ContainerFlexNoWrapDiv>
-                <GrowingDiv>
-                  <HeadingDiv>Start timestamp of energy use</HeadingDiv>
-                  <InputWrapDiv>
-                    <InputHalfWrap>
-                      <IconWrap>
-                        <IconDate></IconDate>Date
-                      </IconWrap>
-                      <Text>{start.split('T')[0].split('-').join('/')}</Text>
-                    </InputHalfWrap>
-                    <InputHalfWrap>
-                      <IconWrap>
-                        <IconTime></IconTime>Time
-                      </IconWrap>
-                      <Text>{start.split('T')[1].split('.')[0]}</Text>
-                    </InputHalfWrap>
-                  </InputWrapDiv>
-                </GrowingDiv>
-                <CertificateTimeInterval
-                  sTimestamp={`${start}`}
-                  eTimestamp={`${end}`}
-                />
-                <GrowingDiv>
-                  <HeadingDiv>End timestamp of energy use</HeadingDiv>
-                  <InputWrapDiv>
-                    <InputHalfWrap>
-                      <IconWrap>
-                        <IconDate></IconDate>Date
-                      </IconWrap>
-                      <Text>{end.split('T')[0].split('-').join('/')}</Text>
-                    </InputHalfWrap>
-                    <InputHalfWrap>
-                      <IconWrap>
-                        <IconTime></IconTime>Time
-                      </IconWrap>
-                      <Text>{end.split('T')[1].split('.')[0]}</Text>
-                    </InputHalfWrap>
-                  </InputWrapDiv>
-                </GrowingDiv>
-              </ContainerFlexNoWrapDiv>
-            </FlexLargeDiv>
-          )}
-          {energy && (
-            <FlexDiv>
-              <ContainerFullWidthWrapDiv>
-                <HeadingDiv>Electric energy use</HeadingDiv>
-                <InputWrapShortDiv>{energy / 1000000} kWh</InputWrapShortDiv>
-              </ContainerFullWidthWrapDiv>
-            </FlexDiv>
-          )}
-          {size && (
-            <FlexDiv>
-              <ContainerFullWidthWrapDiv>
-                <HeadingDiv>H2 batch size</HeadingDiv>
-                <InputWrapShortDiv>{size / 1000000} kWh</InputWrapShortDiv>
-              </ContainerFullWidthWrapDiv>
-            </FlexDiv>
-          )}
-          {eco2 && (
-            <FlexRoundedDiv>
-              <ContainerFullWidthWrapDiv>
-                <HeadingDiv>Carbon Embodiment</HeadingDiv>
-                <InputWrapDiv>{eco2} g CO2e</InputWrapDiv>
-              </ContainerFullWidthWrapDiv>
-            </FlexRoundedDiv>
-          )}
-        </ContainerFlexWrapDiv>
-      </PaddedWrapperDiv>
-    </>
-  )
-}
+// import BgIconTimeSVG from '../assets/images/icon-time.svg'
 
 export default function CertificateViewer() {
   const { current } = useContext(Context)
@@ -265,7 +43,7 @@ export default function CertificateViewer() {
       alert(JSON.stringify(e))
     }
     if (JSON.stringify(res) != JSON.stringify(data)) setData(res)
-    console.log(Math.random())
+    // console.log(Math.random())
   }, [origin, id, data, callApiFn])
 
   // Query API every two seconds
